+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 # Change Log
 
 ## [In Development](https://github.com/ppfeufer/eve-online-intel-tool/tree/development)
-[Full Changelog](https://github.com/ppfeufer/eve-online-intel-tool/compare/v0.1.7...HEAD)
+[Full Changelog](https://github.com/ppfeufer/eve-online-intel-tool/compare/v0.1.8...HEAD)
+- in development
+
+## [v0.1.8](https://github.com/ppfeufer/eve-online-intel-tool/releases/tag/v0.1.8) - 2017-09-02
+[Full Changelog](https://github.com/ppfeufer/eve-online-intel-tool/compare/v0.1.7...v0.1.8)
 ### Added
 - Visual association of pilots and corporations to their respective alliance in local and fleet composition scans